Update BallClassification1.py
This is Simple Ball Classification Case Study which is used for learn Concepts.In this I can used sklearn library which is most powerful and used for classification , regression. And this case study I used Decision Tree Classifier .This is classify data in Yes or No format.
Example:- If I got mail which is spam or not to check we can used this decision tree classifier algorithm.